+++ compute_space/src/compute_space/web/routes/pages/login.py
@@ -12,7 +12,6 @@ from quart import request
 from quart import url_for
 from quart.typing import ResponseReturnValue
 
-from compute_space.core.auth.auth import get_current_user_from_request
 from compute_space.core.auth.tokens import REFRESH_TOKEN_EXPIRY
 from compute_space.core.auth.tokens import create_access_token
 from compute_space.db import get_db
@@ -20,8 +19,7 @@ from compute_space.web.auth.cookies import COOKIE_ACCESS
 from compute_space.web.auth.cookies import COOKIE_REFRESH
 from compute_space.web.auth.cookies import clear_auth_cookies
 from compute_space.web.auth.cookies import set_auth_cookies
-from compute_space.web.auth.middleware import _try_refresh  # noqa: F401 — re-exported
-from compute_space.web.auth.middleware import login_required  # noqa: F401 — re-exported
+from compute_space.web.auth.middleware import get_current_user_from_request
 
 auth_bp = Blueprint("auth", __name__)
